doc: add comments for school work purposes

+++ Service/PurchaseService.cpp
@@ -8,7 +8,6 @@
 #include "DataAccess/IProductRepository.h"
 
 namespace Service {
-
     PurchaseService::PurchaseService(DataAccess::ICartOrderRepository &cartOrderRepository,
                                      DataAccess::ICouponRepository &couponRepository)
             : cartOrderRepository(cartOrderRepository), couponRepository(couponRepository) {
@@ -17,6 +16,7 @@ namespace Service {
 
     PurchaseResult PurchaseService::purchase(const std::vector<std::pair<DataType::FullProduct, long>> &productList,
                                              const std::optional<DataType::Coupon> &coupon) {
+        // 先检测库存是否足够
         DataAccess::IProductRepository &productRepository = StorageService::getInstance()->getProductRepository();
         for (const auto &product: productList) {
             // Check is stock is enough
@@ -25,9 +25,10 @@ namespace Service {
             }
         }
 
-        // now we can proceed with the purchase
+        // 开始进行购买，先将订单加入订单表
         cartOrderRepository.addOrder(productList, calculateTotalPrice(productList, coupon));
 
+        // 更新库存，添加商品到订单表，删除购物车中的商品
         for (const auto &product: productList) {
             cartOrderRepository.removeProduct(product.first.getId());
             DataType::FullProduct newProduct = productRepository.getProduct(product.first.getId()).value();
@@ -35,9 +36,12 @@ namespace Service {
             productRepository.updateProduct(newProduct);
         }
 
+        // 如果使用了优惠券，则将优惠券从用户中删除
         if (coupon.has_value()) {
-            couponRepository.removeCouponFromUser(SessionManager::getInstance()->getCurrentUser().value().getUsername(),
-                                                  coupon.value().getId());
+            couponRepository.removeCouponFromUser(
+                    SessionManager::getInstance()->getCurrentUser().value().getUsername(),
+                    coupon.value().getId()
+            );
         }
 
         return PurchaseResult::success();
@@ -49,14 +53,21 @@ namespace Service {
         for (const auto &product: productList) {
             totalPrice += product.first.getActualPrice() * static_cast<double>(product.second);
         }
+        // 如果有优惠券，则应用优惠券
         if (coupon.has_value()) {
             totalPrice = coupon.value().apply(totalPrice);
         }
         return totalPrice;
     }
 
+    // 购买单个商品，只需要处理库存问题
     PurchaseResult PurchaseService::purchase(const DataType::FullProduct &product) {
         DataAccess::IProductRepository &productRepository = StorageService::getInstance()->getProductRepository();
+
+        if (productRepository.getProduct(product.getId()).value().getRemainingStock() < 1) {
+            return PurchaseResult::stockNotEnough(product.getName());
+        }
+
         cartOrderRepository.removeProduct(product.getId());
         DataType::FullProduct newProduct = productRepository.getProduct(product.getId()).value();
         newProduct.setRemainingStock(newProduct.getRemainingStock() - 1);
